Updated to VS 2019, .Net Core 2.1 and Markdown
