--- Serverless/Serverless App with Dynamo-Example/TestServerless.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2002
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestServerless", "TestServerless\TestServerless.csproj", "{7C9D0F7D-1C02-4D1D-A679-B16146E70CBA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7C9D0F7D-1C02-4D1D-A679-B16146E70CBA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C9D0F7D-1C02-4D1D-A679-B16146E70CBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C9D0F7D-1C02-4D1D-A679-B16146E70CBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C9D0F7D-1C02-4D1D-A679-B16146E70CBA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FD878279-5366-4E87-B59D-E4F4AF100B81}
-	EndGlobalSection
-EndGlobal
