--- Serverless/Serverless App with Dynamo-Example/TestServerless/Function.cs
@@ -1,51 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading.Tasks;
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.Model;
-using Newtonsoft.Json;
-using Amazon.Lambda.Core;
-using Amazon.Lambda.APIGatewayEvents;
-using Amazon.DynamoDBv2.DataModel;
-
-// Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
-[assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.Json.JsonSerializer))]
-
-namespace TestServerless
-{
-    public class Functions
-    {
-        IDynamoDBContext DDBContext { get; set; }
-        /// <summary>
-        /// Default constructor that Lambda will invoke.
-        /// </summary>
-        public Functions()
-        {
-            this.DDBContext = new DynamoDBContext(new AmazonDynamoDBClient());
-        }
-
-
-        /// <summary>
-        /// A Lambda function to respond to HTTP Get methods from API Gateway
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns>The list of blogs</returns>
-        public async Task<APIGatewayProxyResponse> GetAsync(APIGatewayProxyRequest request, ILambdaContext context)
-        {
-            context.Logger.LogLine("Get Request\n");
-
-            var search = this.DDBContext.ScanAsync<Book>(null);
-            var page = await search.GetNextSetAsync();
-            var response = new APIGatewayProxyResponse
-            {
-                StatusCode = (int)HttpStatusCode.OK,
-                Body = JsonConvert.SerializeObject(page),
-                Headers = new Dictionary<string, string> { { "Content-Type", "application/json" } }
-            };
-
-            return response;
-        }
-    }
-}
