+++ Serverless/Serverless App with Dynamo-Example/TestServerless/TestServerless.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29230.47
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestServerless", "TestServerless\TestServerless.csproj", "{7818B2B6-59C9-4A00-B8AF-EC56DA5CE148}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7818B2B6-59C9-4A00-B8AF-EC56DA5CE148}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7818B2B6-59C9-4A00-B8AF-EC56DA5CE148}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7818B2B6-59C9-4A00-B8AF-EC56DA5CE148}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7818B2B6-59C9-4A00-B8AF-EC56DA5CE148}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {AC00E91C-A2C0-4D8E-AF40-AB5942988B5B}
+	EndGlobalSection
+EndGlobal
