--- Serverless/Serverless App with Dynamo-Example/TestServerless/aws-lambda-tools-defaults.json
@@ -1,19 +1,0 @@
-
-{
-    "Information" : [
-        "This file provides default values for the deployment wizard inside Visual Studio and the AWS Lambda commands added to the .NET Core CLI.",
-        "To learn more about the Lambda commands with the .NET Core CLI execute the following command at the command line in the project root directory.",
-        "dotnet lambda help",
-        "All the command line options for the Lambda command can be specified in this file."
-    ],
-    "profile"     : "aws-internal",
-    "region"      : "us-east-1",
-    "configuration" : "Release",
-    "framework"     : "netcoreapp2.0",
-    "s3-prefix"     : "TestServerless/",
-    "template"      : "serverless.template",
-    "template-parameters" : "",
-    "s3-bucket"           : "serverlessapptest",
-    "stack-name"          : "TestServerless",
-    "stack-wait"          : false
-}
