--- Serverless/Serverless App with Dynamo-Example/TestServerless/serverless.template
@@ -1,65 +1,0 @@
-{
-  "AWSTemplateFormatVersion" : "2010-09-09",
-  "Transform" : "AWS::Serverless-2016-10-31",
-  "Description" : "An AWS Serverless Application.",
-
-  "Resources" : {
-
-    "Get" : {
-      "Type" : "AWS::Serverless::Function",
-      "Properties": {
-        "Handler": "TestServerless::TestServerless.Functions::GetAsync",
-		"FunctionName": "TestServerless",
-        "Runtime": "dotnetcore2.0",
-        "CodeUri": "",
-        "MemorySize": 128,
-        "Timeout": 30,
-        "Role": null,
-        "Policies": [ "AWSLambdaFullAccess" ],
-        "Events": {
-          "PutResource": {
-            "Type": "Api",
-            "Properties": {
-              "Path": "/",
-              "Method": "GET"
-            }
-          }
-        }
-      }
-    },
-  "readingListDynamoDBTable" : {
-      "Type" : "AWS::DynamoDB::Table",
-      "Properties" : {
-        "AttributeDefinitions" : [
-          {
-            "AttributeName" : "ItemId",
-            "AttributeType" : "S"   
-          },
-          {
-            "AttributeName" : "Title",
-            "AttributeType" : "S"
-          }
-        ],
-        "KeySchema" : [
-          {
-            "AttributeName" : "ItemId",
-            "KeyType" : "HASH"
-          },
-		  {
-            "AttributeName" : "Title",
-            "KeyType" : "RANGE"
-          }
-        ],
-        "ProvisionedThroughput" : {
-          "ReadCapacityUnits" : "1",
-          "WriteCapacityUnits" : "1"
-        },
-        "TableName" : "readingList"
-	  }
-	}
-  },
-
-
-  "Outputs" : {
-  }
-}
